added the optional --update-pythons parameter to installation script improved pyegi dependencies installation fixed symlink prevelege issue exclude normalization in rmtree
