+++ src/dependency/Pyegi/Installer/installer.py
@@ -12,8 +12,7 @@ from Pyegi.minimals.minimal_utils import (
     VenvEnvBuilder,
 )
 from Pyegi.minimals.minimal_installer import (
-    commonize_venv,
-    infer_python_version,
+    install_pkgs,
     pyproject_file,
     poetry_toml_file,
 )
@@ -27,6 +26,7 @@ def _create_env(parent_path, update=False):
 if __name__ == "__main__":
     parser = argparse.ArgumentParser()
     parser.add_argument("--install", action="store_true")
+    parser.add_argument("--update-pythons", action="store_true")
     parser.add_argument("--venv", type=str)
     parser.add_argument("--update", action="store_true")
     args, unknown_args = parser.parse_known_args()
@@ -44,19 +44,20 @@ if __name__ == "__main__":
         shutil.rmtree(
             normal_path_join(automation_path, "include", "Pyegi"), ignore_errors=True
         )
-        rmtree(dependency_dir, ["Pyegi/settings.json", "Pythons/common-packages"])
+        excludes = ["Pyegi/settings.json"]
+        if args.update_pythons:
+            excludes.append("Pythons/common-packages")
+        else:
+            excludes.append("Pythons")
+        rmtree(dependency_dir, excludes)
         # copy new files
+        # NOTICE: scripts' binaries are broken because of the static linking
         shutil.copytree("src", automation_path, dirs_exist_ok=True)
         shutil.copy(poetry_toml_file, pyegi_dir)
         shutil.copy(pyproject_file, pyegi_dir)
 
         # install Pyegi dependencies
-        os.chdir(pyegi_dir)
-        py_version = infer_python_version(pyegi_dir + pyproject_file).python_version
-        py_version.create_env(pyegi_dir)
-        # NOTICE: scripts' binaries are broken because of the static linking
-        result = py_version.run_module(["poetry", "install", "--no-dev"])
-        commonize_venv(pyegi_dir)
+        install_pkgs(pyegi_dir)
     elif args.venv != None:
         if len(args.venv) == 0:
             raise ValueError("You didn't provide any path for venv's parent!")
